Removed aliasing of primary_key_name to foreign_key
Prior to this commit, I was getting this message repeatedly during an infinite recursion:

  DEPRECATION WARNING: primary_key_name is deprecated and will be removed from Rails 3.2 (use foreign_key instead). (called from block in destroy at /Users/craig/Sites/rails3_acts_as_paranoid/lib/rails3_acts_as_paranoid.rb:125)

...and then the StackTraceToDeep error, if I let it get that far.

I didn't dig to deep into it, but it looks like goncalossilva et al
had roughly the same idea as the Rails 3.1 team, and made primary_key_name
call foreign_key; with the alias this caused an infinite recursion.

In any case, all of the tests pass fine with this alias removed.